Feature/scan 296 analytics stats methods

+++ apps/web-api/src/statistics/statistics.dto.ts
@@ -1,11 +1,11 @@
 import { Total } from '@entities/Total';
-import { Field, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, ObjectType } from '@nestjs/graphql';
 
 @ObjectType('statistics')
 export class StatisticsDTO implements Partial<Total> {
   @Field(() => String, { nullable: true })
   name?: string;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   count?: number;
 }
